Move platforms.h back out of shared folder
Although platforms.h is shared it makes sense to be in the HAL folder as it controls header redirects and needs updated when a new platform folder is added
