+++ Marlin/src/inc/MarlinConfigPre.h
@@ -23,7 +23,7 @@
 #ifndef _MARLIN_CONFIGPRE_H_
 #define _MARLIN_CONFIGPRE_H_
 
-#include "../HAL/shared/platforms.h"
+#include "../HAL/platforms.h"
 #include "../core/boards.h"
 #include "../core/macros.h"
 #include "../core/types.h"
